+++ src/main/java/sayner/sandbox/controllers/UserController.java
@@ -16,7 +16,7 @@ import sayner.sandbox.dto.mappers.UserMapper;
 import sayner.sandbox.dto.status.enums.StatusEnum;
 import sayner.sandbox.dto.views.SingleResponseObjectDtoView;
 import sayner.sandbox.model.enums.RoleEnum;
-import sayner.sandbox.services.UserService;
+import sayner.sandbox.services.CurrentUserDetailsService;
 
 @RequiredArgsConstructor(onConstructor = @__({@Autowired}))
 @RestController
@@ -27,7 +27,8 @@ public class UserController {
     final String GODLiKE = RoleEnum.ROLE_GODLiKE.name();
 
     private final UserMapper userMapper = UserMapper.INSTANCE;
-    private final UserService userService;
+
+    private final CurrentUserDetailsService userService;
 
     @GetMapping
     @JsonView(SingleResponseObjectDtoView.FullWithUserFull.class)
