+++ src/main/java/sayner/sandbox/services/CurrentUserDetailsService.java
@@ -2,5 +2,5 @@ package sayner.sandbox.services;
 
 import org.springframework.security.core.userdetails.UserDetailsService;
 
-public interface CustomUserDetailsService extends UserDetailsService {
+public interface CurrentUserDetailsService extends UserService, UserDetailsService {
 }
