+++ src/main/java/sayner/sandbox/services/impl/UserServiceImpl.java
@@ -1,7 +1,6 @@
 package sayner.sandbox.services.impl;
 
 import lombok.extern.log4j.Log4j2;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 import sayner.sandbox.model.User;
 import sayner.sandbox.services.UserService;
@@ -10,7 +9,6 @@ import java.util.Arrays;
 import java.util.List;
 
 @Service
-@Primary
 @Log4j2
 public class UserServiceImpl implements UserService {
